gdbstub: Use size_t for strlen() return value
Since strlen() returns an unsigned value, it is pointless to
convert it to a signed one. Use size_t to hold its return value.

Message-Id: <20190220010232.18731-6-philmd@redhat.com>

+++ gdbstub.c
@@ -1693,7 +1693,7 @@ static int gdb_handle_packet(GDBState *s, const char *line_buf)
         }
 #else /* !CONFIG_USER_ONLY */
         else if (strncmp(p, "Rcmd,", 5) == 0) {
-            int len = strlen(p + 5);
+            size_t len = strlen(p + 5);
 
             if ((len % 2) != 0) {
                 put_packet(s, "E01");
